requirements.txt
# Core Dependencies
streamlit==1.32.0
pandas==2.1.0
numpy==1.26.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
pillow==10.1.0
opencv-python==4.8.1.78

# Email Dependencies
yagmail==0.15.293

# Web Scraping Dependencies
lxml==4.9.3
html5lib==1.1

# Google Search D
googlesearch-python==1.2.3

# WhatsApp Messaging
pywhatkit==5.4

# Instagram Integration
instagrapi==1.16.22

# Google Generative AI
google-generativeai==0.3.2

# PDF Generation
fpdf2==2.7.8

# Agentic AI Assistant
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
langgraph>=0.0.15
wikipedia>=1.4.0
speechrecognition>=3.10.0
langchain>=0.0.300
gradio>=3.50.0
python-magic>=0.4.27